3_semester_2024/prog/lab/lista1/src/main/java/org/rafisto/Tourist.java
package org.rafisto;

public final class Tourist {
    private String name;
    private String surname;
    private double money;

    public Tourist(String name, String surname, double money) {
        if (name == null || name.isEmpty()) {
            throw new IllegalArgumentException("Name must not be null or empty");
        }

        if (surname == null || surname.isEmpty()) {
            throw new IllegalArgumentException("Surname must not be null or empty");
        }
        this.name = name;
        this.surname = surname;
        this.money = money;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) throws IllegalArgumentException {
        if (name == null || name.isEmpty()) {
            throw new IllegalArgumentException("Name must not be null or empty");
        }
        this.name = name;
    }

    public String getSurname() {
        return surname;
    }

    public void setSurname(String surname) throws IllegalArgumentException {
        if (surname == null || surname.isEmpty()) {
            throw new IllegalArgumentException("Surname must not be null or empty");
        }
        this.surname = surname;
    }

    public double getMoney() {
        return money;
    }

    public void setMoney(double money) {
        this.money = money;
    }
}
